working on the object feedback
Fist I tried to make the distance work itself out in their own
functions, but then I seem like a lot of function activation for
something that could easily be down in the encounter function.
I have therefor made a separat function. This way we will return the
object to poets factual encounter, leading to the pet being able to
decide on its own if it wants to interact with the object or not.