Upgrading the smell locating
Improved the smell locating feature. Also moved the "exploring" done by
the pet inside the move function in the pet class.
The smell feature, now generates a random target witin the smell range
on the side the pet comes from. The pet will then get to cero in with
random tries.
It has to efectively -sniff- out the target object.